Implement support for code-backed migrations.
This wraps the normal SQL migration steps taken by the program in some glue code, enabling us to ship
additional complex operations side-by-side with the SQL code. Arbitrary Rust code can be inserted before and/or
after the up and down stages of a migration, giving us the chance to act on the database in both its old
and new schema.

Code migrations currently need to be registered manually in the get_code_migration function, mapping to diesel's generated migration name.